changes + comments

+++ match.go
@@ -7,16 +7,16 @@ import (
 
 // Match represents a matched pattern in the input.
 type Match struct {
-	pos     int64
-	pattern int64
+	pos     uint32
+	pattern uint32
 	match   []byte
 }
 
-func newMatch(pos, pattern int64, match []byte) *Match {
+func newMatch(pos, pattern uint32, match []byte) *Match {
 	return &Match{pos, pattern, match}
 }
 
-func newMatchString(pos, pattern int64, match string) *Match {
+func newMatchString(pos, pattern uint32, match string) *Match {
 	return &Match{pos: pos, pattern: pattern, match: []byte(match)}
 }
 
@@ -25,24 +25,16 @@ func (m *Match) String() string {
 }
 
 // Pos returns the byte position of the match.
-func (m *Match) Pos() int64 {
-	return m.pos
-}
+func (m *Match) Pos() uint32 { return m.pos }
 
 // Pattern returns the pattern id of the match.
-func (m *Match) Pattern() int64 {
-	return m.pattern
-}
+func (m *Match) Pattern() uint32 { return m.pattern }
 
 // Match returns the pattern matched.
-func (m *Match) Match() []byte {
-	return m.match
-}
+func (m *Match) Match() []byte { return m.match }
 
 // MatchString returns the pattern matched as a string.
-func (m *Match) MatchString() string {
-	return string(m.match)
-}
+func (m *Match) MatchString() string { return string(m.match) }
 
 // MatchEqual check whether two matches are equal (i.e. at same position, pattern and same pattern).
 func MatchEqual(a, b *Match) bool {
